Move shell call related mocks to piper_test.go
